NPE when stopping puller
- Added a null check before stopping pull replication in order to
  avoid NPEs
- Added some logs
- Downgraded the gradle version because of issues with building the project

+++ src/main/java/com/couchbase/lite/replicator/Puller.java
@@ -223,8 +223,15 @@ public final class Puller extends Replication implements ChangeTrackerClient {
     @InterfaceAudience.Private
     protected void stopped() {
         //TODO verify it this doesn't affects cbl logic
-        downloadsToInsert.flushAll();
-        downloadsToInsert = null;
+
+        if (downloadsToInsert != null) {
+            Log.w(Log.TAG_SYNC, "Puller.stopped(): downloadsToInsert is null!");
+            downloadsToInsert.flushAll();
+            downloadsToInsert = null;
+        } else {
+            Log.w(Log.TAG_SYNC, "Puller.stopped(): downloadsToInsert is not null!");
+        }
+
         super.stopped();
     }
